block more revisions ...
TO ALL DEVELOPERS WITH COMMIT ACCESS:
Please remember to *always* do an 'svnmerge block -r <rev#>' on the trunk
every single time you make a commit to the 1.2 branch which will not be pulled
into the trunk using svnmerge.  This is important for us to be able to take
full advatage of this tool.


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@8089 f38db490-d61c-443f-a65b-d21fe96a405b