Blocked revisions 41411 via svnmerge
........
r41411 | russell | 2006-08-30 14:59:44 -0400 (Wed, 30 Aug 2006) | 6 lines

Restore original functionality of 1.2 in places where ANI was not set, but was
changed to be set.  The original change was done to ensure that the behavior of
the "callerid" option in each channel driver was consistent, but it caused an
unexpected behavior change of CDR records for users, so this change is being
reverted in 1.2.  (issue #7695)

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@41429 f38db490-d61c-443f-a65b-d21fe96a405b