Updating sip.conf.sample with information about T38 not working when chan_local or chan_agent is involved in the call.
I don't know how big a fix that would be to solve, but this is
the current state of affairs.

(Chan_sip currently checks if the other side of the bridge
has a SIP tech. We could/should implement another check,
possibly for udptl_write or some flag in the ast_channel
structure).


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@48264 f38db490-d61c-443f-a65b-d21fe96a405b