Blocked revisions 55750 via svnmerge
........
r55750 | russell | 2007-02-20 18:19:14 -0600 (Tue, 20 Feb 2007) | 9 lines

Fix random crashes when using the MeetMe application.  This patch converts list
handling to use the linked list macros and most importantly, implements
reference counting on the ast_conference objects.  The reference counting was
first backported from 1.4.  However, that code has some problems that caused
the reference count to never hit zero.  Those problems are fixed in this patch
and will be resolved in 1.4 and trunk next, with a different patch.

(issues #7647, #9073, #9106, BE-115). 

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@55751 f38db490-d61c-443f-a65b-d21fe96a405b