Merged revisions 60134 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.2
........
r60134 | russell | 2007-04-04 12:38:47 -0500 (Wed, 04 Apr 2007) | 6 lines

It is valid to redirect channels via the manager interface that are not in the
UP state.  Instead of checking for that to prevent to ensure a dead channel
doesn't get redirected, just use the ast_check_hangup() API call.
(issue #9457, reported by Callmewind, patch by me)
(related to issue #8977)

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@60137 f38db490-d61c-443f-a65b-d21fe96a405b