(closes issue #10301) Reported by: fnordian Patches:       asterisk-1.4.9-channel.c.patch uploaded by fnordian (license 110)       Additional changes by me
Fix some problems in channel_find_locked() which can cause an infinite loop.
The reference to the previous channel is set to NULL in some cases.  These changes
ensure that the reference to the previous channel gets restored before needing
it again.

I'm not convinced that the code that is setting it to NULL is really the right
thing to do.  However, I am making these changes to fix the obvious problem
and just leaving an XXX comment that it needs a better explanation that what
is there now.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@77780 f38db490-d61c-443f-a65b-d21fe96a405b