Merged revisions 78749 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r78749 | russell | 2007-08-09 12:24:40 -0500 (Thu, 09 Aug 2007) | 9 lines

Fix subscriptions to multiple mailboxes for ODBC_STORAGE.  Also, leave a
comment for this to be fixed for IMAP_STORAGE, as well.  I left IMAP alone
since I know MarkM was working on this code right now for another reason.

This is broken even worse in trunk, but for a different reason.  The fact
that the mailbox option supported multiple mailboxes is completely not obvious
from the code in the channel drivers.  Anyway, I will fix that in another
commit ...

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@78750 f38db490-d61c-443f-a65b-d21fe96a405b