Merged revisions 81416 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r81416 | mmichelson | 2007-08-31 14:48:55 -0500 (Fri, 31 Aug 2007) | 6 lines

Fixed broken behavior of a reload on realtime queues. Prior to this patch, if a reload was issued and
a realtime queue had callers waiting in it, then the queue would be removed from the queue list, but it would
not actually be freed (in fact, a debug message warning about a memory leak would come up). With this patch,
reloads do not touch realtime queues at all.


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@81417 f38db490-d61c-443f-a65b-d21fe96a405b