Change the traversal to use ao2_callback() instead of an ao2_iterator.  Using ao2_callback() is a much more efficient way of performing an operation on every item in the container.  This change makes hashtest2 run in about 25% of the time it ran before on my system.
In general, I would say that it makes the most sense to use an ao2_iterator if
the operation being performed is going to take a long time and you don't want
to keep the container locked while you work with each object.  Otherwise,
the use of ao2_callback is preferred.


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@82282 f38db490-d61c-443f-a65b-d21fe96a405b