Merged revisions 86471 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r86471 | file | 2007-10-19 12:33:49 -0300 (Fri, 19 Oct 2007) | 6 lines

Fix two issues with domains and transfers. If a port was given in the hostname it was treated as part of the hostname. If domains were configured but external domains were not enabled all transfers would be considered remote.
(closes issue #11027)
Reported by: ramonpeek
Patches:
      11027-1.diff uploaded by ramonpeek (license 266)

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@86472 f38db490-d61c-443f-a65b-d21fe96a405b