Merged revisions 90470 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r90470 | russell | 2007-12-02 12:18:52 -0600 (Sun, 02 Dec 2007) | 6 lines

The other day when I went through making changes as a result of the ao2_link()
change, I added some code to set pointers to NULL after they were unreferenced.
This pointed out that in this place, the object was unreferenced before the
code was done using it.  So, move the unref down a little bit.
(crash reported by jmls on IRC)

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@90471 f38db490-d61c-443f-a65b-d21fe96a405b