Avoid reinvite race situations with two Asterisks trying to reinvite each other in 1.4 and trunk.
This patch implements support for the 491 error code that
Asterisk 1.4 generates on situations where we get an 
incoming INVITE and already has one in progress.

Thanks to mavetju for reporting and to Raj Jain for an
excellent explanation of the problem.

Patch by myself. Tested with 8 Asterisk servers connected
to each other in a training network.

Closes issue #10481



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@92158 f38db490-d61c-443f-a65b-d21fe96a405b