Merged revisions 95890 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r95890 | mmichelson | 2008-01-02 11:51:22 -0600 (Wed, 02 Jan 2008) | 9 lines

A change to improve the accuracy of queue logging in the case where a member does not
answer during the specified timeout period. Prior to this change, there was a small chance
that the member name recorded in this case would be blank. Also prior to this change, if using
the ringall strategy, if no one answered the call during the specified timeout, the member name
listed in the queue log would randomly be one of the members that was rung.

(closes issue #11498, reported and tested by hloubser, patched by me)


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@95891 f38db490-d61c-443f-a65b-d21fe96a405b