Final round of changes for configure script logic for IMAP
Now if a directory is specified, then we will search that directory for
a source installation of the IMAP toolkit. If none is found, then we will
use that directory as the basis for detecting a package installation of
the IMAP c-client. If that check fails, then configure will fail.



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@103722 f38db490-d61c-443f-a65b-d21fe96a405b