Merged revisions 110303 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r110303 | russell | 2008-03-20 15:08:26 -0500 (Thu, 20 Mar 2008) | 8 lines

Fix a bug when using zaptel timing for playing back files that have a sample rate
other than 8 kHz.  The issue here is that format modules give a "whennext" sample
value, which is used to calculate when to set a timer for to retrieve the next
frame.  However, the zaptel timer operates on 8 kHz samples, so this must be taken
into account.

(another part of issue #12164, reported by milazzo and jsmith, patch by me)

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@110304 f38db490-d61c-443f-a65b-d21fe96a405b