Blocked revisions 114049 via svnmerge
........
r114049 | file | 2008-04-10 17:28:40 -0300 (Thu, 10 Apr 2008) | 2 lines

A 'b' option has been added which causes chan_local to return the actual channel that is behind it when queried. This is useful for transfer scenarios as the actual channel will be transferred, not the Local channel. If you have been using Local channels as queue members and having issues when the agent did a blind transfer this option may solve the issue.

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@114050 f38db490-d61c-443f-a65b-d21fe96a405b