Merged revisions 114327 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r114327 | jpeeler | 2008-04-21 10:34:37 -0500 (Mon, 21 Apr 2008) | 2 lines

This removes an invalid warning message for an incorrectly entered pin, but more importantly removes an inapplicable check. If the first argument passed to app_authenticate does not contain a '/', the argument should be treated as the sole fixed "password" to match against and that is all. (Previous behavior was attempting to open a file based on the pin.)

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@114328 f38db490-d61c-443f-a65b-d21fe96a405b