Merged revisions 114553 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r114553 | murf | 2008-04-22 15:57:57 -0600 (Tue, 22 Apr 2008) | 14 lines


(closes issue #12469)
Reported by: triccyx

I had a bit a problem reproducing this in my setup (trying not to disturb my other stuff)
but finally, I got it. The problem appears to be that the extension is being added in
replace mode, which kinda assumes that the pattern trie has been formed, when in fact,
in this case, it was not. The checks being done are not nec. when the tree is not yet
formed, as changes like this will be summarized when the trie is formed in the future.

I tested the fix, and the crash no longer happens. Feel free to open the bug again if
this fix doesn't cure the problem.


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@114556 f38db490-d61c-443f-a65b-d21fe96a405b