Blocked revisions 114834 via svnmerge
........
r114834 | mmichelson | 2008-04-29 12:56:13 -0500 (Tue, 29 Apr 2008) | 8 lines

Since there is now a globally available function for saying someone's name, a LOT of
functions in app_directory can be removed since the ODBC-specific lookups are accomplished
within app_voicemail. This change greatly reduces the amount of lines in app_directory that
were solely for the purpose of looking up a name when ODBC_STORAGE is specified for voicemail.

This commit also makes the name-saying interruptable via DTMF.


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@114835 f38db490-d61c-443f-a65b-d21fe96a405b