Blocked revisions 114857 via svnmerge
........
r114857 | mmichelson | 2008-04-29 16:07:36 -0500 (Tue, 29 Apr 2008) | 15 lines

Patching app_chanspy to jibe better with what is documented. This allows for 
a colon-delimited list of spygroups to be specified when calling the ChanSpy application
with the 'g' option. Prior to this, you could only specify a single group when using the
'g' option.

I also have upped the maximum number of spygroups to 128 and added a #define so that this
can be easily increased or decreased later.

(closes issue #12497)
Reported by: jsmith
Patches:
      app_chanspy_multiple_groups_v2.patch uploaded by jsmith (license 15)
Tested by: atis, jvandal


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@114858 f38db490-d61c-443f-a65b-d21fe96a405b