Merged revisions 115197 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
................
r115197 | mmichelson | 2008-05-02 09:28:55 -0500 (Fri, 02 May 2008) | 14 lines

Merged revisions 115196 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
r115196 | mmichelson | 2008-05-02 09:28:19 -0500 (Fri, 02 May 2008) | 6 lines

Clarify a comment that was, well, just wrong. It turns out that
ignoring the way that macros expand. Instead, I have clarified in the
comment why the macro will work even if the scheduler id for the
task to be deleted changes during the execution of the macro.


........

................


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@115198 f38db490-d61c-443f-a65b-d21fe96a405b