Merged revisions 115321 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
................
r115321 | mmichelson | 2008-05-05 16:43:21 -0500 (Mon, 05 May 2008) | 21 lines

Merged revisions 115320 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
r115320 | mmichelson | 2008-05-05 16:41:34 -0500 (Mon, 05 May 2008) | 13 lines

Don't consider a caller "handled" until the caller is bridged with
a queue member. There was too much of an opportunity for the member
to hang up (either during a delay, announcement, or overly long
agi) between the time that he answered the phone and the time when
he actually was bridged with the caller. The consequence of this
was that if the member hung up in that interval, then proper
abandonment details would not be noted in the queue log if the caller
were to hang up at any point after the member hangup.

(closes issue #12561)
Reported by: ablackthorn


........

................


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@115322 f38db490-d61c-443f-a65b-d21fe96a405b