Blocked revisions 115537 via svnmerge
........
r115537 | russell | 2008-05-07 16:11:33 -0500 (Wed, 07 May 2008) | 10 lines

Fix up a problem that was introduced into the scheduler when it was converted
to use doubly linked lists.  The schedule() function had an optimization that
had it try to guess which direction would be better for the traversal to insert
the task into the scheduler queue.  However, if the code chose the path where
it traversed the queue in reverse, and the result was that the task should be
at the head of the queue, then the code would actually put it at the tail,
instead.

(Problem found by bbryant, debugged and fixed by bbryant and me)

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@115538 f38db490-d61c-443f-a65b-d21fe96a405b