Merged revisions 120171 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r120171 | tilghman | 2008-06-03 17:05:16 -0500 (Tue, 03 Jun 2008) | 5 lines

Move compatibility options into asterisk.conf, default them to on for upgrades,
and off for new installations.  This includes the translation from pipes to commas
for pbx_realtime and the EXEC command for AGI, as well as the change to the Set
application not to support multiple variables at once.

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@120172 f38db490-d61c-443f-a65b-d21fe96a405b