Blocked revisions 122228 via svnmerge
........
r122228 | mmichelson | 2008-06-12 11:25:09 -0500 (Thu, 12 Jun 2008) | 12 lines

Merging the work done in the queue-log-atxfer branch. The
net result of this work is that attended transfers made
by queue members will now show up in the queue_log as a 
TRANSFER message instead of COMPLETECALLER as it had been.

As far as the details go, I created a datastore which is
attached to the calling channel just prior to when the caller
is bridged with the queue member. If the calling channel
is masqueraded, then during the "fixup" portion, the TRANSFER
will be logged and the datastore will be removed.


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@122230 f38db490-d61c-443f-a65b-d21fe96a405b