Merged revisions 131243 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
................
r131243 | murf | 2008-07-16 11:59:33 -0600 (Wed, 16 Jul 2008) | 27 lines

Merged revisions 131242 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
r131242 | murf | 2008-07-16 11:53:43 -0600 (Wed, 16 Jul 2008) | 19 lines

(closes issue #13090)
Reported by: murf

The problem was that, esoteric as it is, because the hangerupper
context immediately preceded the std-priv-extent macro, that 
the checking code accidentally would fall from traversing hangerupper
into the std-priv-exten macro, where it would hit the hangerupper
in the 'includes', and proceed into an infinite recursion.

A small fix to traverse into the statements of the context instead
of the context solves this issue.

I also added some commented out printfs for debug, which were pretty
handy in the face of a dorky gdb.

This was a problem around since the package was first written;
but evidently pretty rare in turning up in the field.


........

................


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@131248 f38db490-d61c-443f-a65b-d21fe96a405b