(closes issue #13197) Reported by: pj
(closes issue #13051)
Reported by: pj

This patch substitutes commas in the expr 
supplied to the if () statement, as in
if ( expr ) ...

This solves both the bugs above, and makes
the source symmetric with switch statements,
which were earlier reported to need this sort
of treatment.

I tested this using the examples, both for
the compiler and at run time. Looks good.



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@134652 f38db490-d61c-443f-a65b-d21fe96a405b