Scrap the 500 ms delay when Asterisk auto-answers a channel. Instead, poll the channel until receiving a voice frame. The cap on this poll is 500 ms.
The optional delay is still allowable in the Answer() application,
but the delay has been moved back to its original position, after
the call to the channel's answer callback. The poll for the voice
frame will not happen if a delay is specified when calling Answer().

(closes issue #12708)
Reported by: kactus



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@136631 f38db490-d61c-443f-a65b-d21fe96a405b