(closes issue #13139) Reported by: krisk84 Tested by: krisk84
This change prevents a call that is placed in the parkinglot to be picked up before the PBX is finished. If another extension dials the parking extension before the PBX thread has completed at minimum warnings will occur about the PBX not properly being terminated. At worst, a crash could occur.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@147941 f38db490-d61c-443f-a65b-d21fe96a405b