Merged revisions 148160 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r148160 | mmichelson | 2008-10-09 18:54:59 -0500 (Thu, 09 Oct 2008) | 14 lines

The priority was unnecessary for the manager atxfer, so it has
been removed. Furthermore, now we actually use the Context argument
passed to set the transfer context and don't error out if no
context is specified.

This addresses the actual problems outlined in issue 12158. Regarding
the other points brought up, regarding the inability to not transfer
to extensions which cannot be represented by DTMF, it is not enough of
a constraint that it is worth attempting to rework the feature.

(closes issue #12158)
Reported by: davidw


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.1@148161 f38db490-d61c-443f-a65b-d21fe96a405b