(closes issue #13891) Reported by: smurfix
This reverts a change I made in 116297. At the time it seemed the change was required to solve an issue with attempting a transfer but then letting it timeout without dialing any digits. However, I didn't realize that having an empty extension was possible. I'm removing the immediate return that was added in pbx_find_extension if the extension is null.



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@156649 f38db490-d61c-443f-a65b-d21fe96a405b