Merged revisions 156649 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r156649 | jpeeler | 2008-11-13 13:17:50 -0600 (Thu, 13 Nov 2008) | 6 lines

(closes issue #13891)
Reported by: smurfix

This reverts a change I made in 116297. At the time it seemed the change was required to solve an issue with attempting a transfer but then letting it timeout without dialing any digits. However, I didn't realize that having an empty extension was possible. I'm removing the immediate return that was added in pbx_find_extension if the extension is null.


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.1@156650 f38db490-d61c-443f-a65b-d21fe96a405b