Blocked revisions 162418 via svnmerge
........
r162418 | russell | 2008-12-09 16:38:41 -0600 (Tue, 09 Dec 2008) | 7 lines

Add some additional Asterisk project developer documentation.  

After the nightly update of the documentation on asterisk.org, I'll post 
an update to asterisk-dev with a pointer to the changes.  This covers some
release branch and commit policy information.  None of this should be a
surprise, since it's just documenting what we have already been doing.

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@162419 f38db490-d61c-443f-a65b-d21fe96a405b