Blocked revisions 167792,167888 via svnmerge
........
r167792 | mmichelson | 2009-01-08 13:48:42 -0600 (Thu, 08 Jan 2009) | 15 lines

Add the average talk time for a queue

This patch adds the functionality to app_queue of calculating
the average amount of time that channels are bridged for a
queue. The algorithm used to calculate the average is the same
exponential average currently used to calculate the average holdtime.
See the CHANGES file to see the methods you may use to view this
information.

(closes issue #13960)
Reported by: coolmig
Patches:
      app_queue.c.diff.trunk-r158840 uploaded by coolmig (license 621)


........
r167888 | mmichelson | 2009-01-08 16:34:52 -0600 (Thu, 08 Jan 2009) | 4 lines

Revert chan_sip changes which were accidentally committed
in revision 167792


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.1@167892 f38db490-d61c-443f-a65b-d21fe96a405b