Make a proper builtin attended transfer to parking work
This is an ugly hack from 1.4 that allows the timeout callback from a parked
call to use the right channel name for the callback when the park is done with
a builtin attended transfer (that isn't completed early).  This hasn't ever
worked in trunk and no one has complained yet, so eh.


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@169510 f38db490-d61c-443f-a65b-d21fe96a405b