Merged revisions 176901 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r176901 | russell | 2009-02-18 00:00:40 -0600 (Wed, 18 Feb 2009) | 9 lines

Fix a number of incorrect uses of strncpy().

The big problem here is that the 3rd argument provided in these uses of strncpy()
did not reserve a byte for the null terminator, leaving the potential for writing
one byte past the end of the buffer.

Aside from this, there were coding guidelines violations with regards to spacing,
as well as hard coded lengths being used instead of sizeof().

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.1@176903 f38db490-d61c-443f-a65b-d21fe96a405b