Merged revisions 186928 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r186928 | russell | 2009-04-08 07:35:57 -0500 (Wed, 08 Apr 2009) | 13 lines

Update some comments and resolve potential memory corruption in chan_sip.

While browsing chan_sip the other day, I noticed this dangerous code in
dialog_needdestroy().  This function is an ao2_callback.  It is absolutely
_not_ okay to unlock the container from within this function.  It's also not
clear why it was useful.  Given that it could cause memory corruption, I have
removed it.

There was also a TODO comment left describing a potential implementation of
an improvement to the needdestroy handling.  I'm not convinced that what was
described is the best choice here, so I have briefly described the way that
this function is used today that could be improved.

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.2@186929 f38db490-d61c-443f-a65b-d21fe96a405b