National prefix inserted even when caller ID not available
When the caller ID is restricted, the expected behavior is for the caller id to be blank.  In chan_dahdi, the national prefix is placed onto the callers number even if its restricted (empty) causing the caller id to be the national prefix rather than blank.

(closes issue #13207)
Reported by: shawkris
Patches:
      national_prefix.diff uploaded by dvossel (license 671)

Review: http://reviewboard.digium.com/r/220/



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@188646 f38db490-d61c-443f-a65b-d21fe96a405b