Make absolute paths for logger channels work properly
(Note: This is not a new feature, it was previously undocumented and broken.)

The Asterisk logger has a feature to support absolute pathnames for logger channels, but the code implementing the feature was broken. This has been fixed, and the absolute path feature is now documented in the sample logger.conf.



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@193193 f38db490-d61c-443f-a65b-d21fe96a405b