Move 300 bytes around on the stack, to make more room for an extension buffer. This allows more concurrent extensions to be copied for a single voicemail, without creating a possibility of upsetting existing users, where a dialplan could run out of stack space where it had run fine before.  Alternatively, we could have allocated off the heap, but that is a larger change and would have increased the chance for instability introduced by this change.
This is really solved starting in 1.6.0.11, as the use of an ast_str buffer
allows an unlimited number of extensions (up to available memory).  We
additionally create a new warning message when the buffer length is exceeded,
permitting administrators to see an issue after the fact, whereas previously
the list was silently truncated.
(closes issue #14739)
 Reported by: p_lindheimer
 Patches: 
       20090417__bug14739.diff.txt uploaded by tilghman (license 14)
 Tested by: p_lindheimer


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@193755 f38db490-d61c-443f-a65b-d21fe96a405b