Removed confusing warning message "Got Busy in Connected State"
If an incoming mISDN call is answered with the Answer application and a
subsequent Dial gets a busy endpoint then it is valid for that already
connected channel to get the busy indication.  Asterisk will play the busy
tones until the dialplan plays something else or hangs up the call.

(closes issue #11974)
Reported by: fvdb


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@204834 f38db490-d61c-443f-a65b-d21fe96a405b