Remove harmful code that causes endless loops. Remove code that causes loops in registrations.
We have agreed that the patch that this code was part of was bad. I am ripping out the code that causes 
the issue. putnopvut needs to check the rest of the patch, if it needs to be changed as well.

This solves the issue reported in #15540, but needs more work before we close it (as described above).



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@217668 f38db490-d61c-443f-a65b-d21fe96a405b