iax2 frame double free
The iax frame's retrans sched id was written over right
before iax2_frame_free was called.  In iax2_frame_free that
retrans id is used to delete the sched item.  By writing over
the retrans field before the sched item could be deleted, it was
possible for a retransmit to occur on a freed frame.



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@219519 f38db490-d61c-443f-a65b-d21fe96a405b