Fix incorrect filename comparsion after monitor file change
The logic to detect if a requested file is indeed a different file from the
current file was incorrect. The main issue being confusion of the use of
filename_base which was previously set without pathing information and then
compared to another full path. Robust file comparison logic has been added
to properly check if two files are the same even if symlinks are used.

(closes issue #15313)
Reported by: caspy
Patches: 
      20091103__issue15313__1.4.diff.txt uploaded by jpeeler (license 325)
      but mostly tilghman's work


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@227944 f38db490-d61c-443f-a65b-d21fe96a405b