Some platforms clear /var/run at boot, which makes connecting a remote console... difficult.
Previously, we only created the default /var/run/asterisk directory at install
time.  While we could create it in the init script, that would not work for
those who start asterisk manually from the command line.  So the safest thing
to do is to create it as part of the Asterisk boot process.  This also changes
the ownership of the directory, because the pid and ctl files are created after
we setuid/setgid.

(closes issue #16802)
 Reported by: Brian
 Patches: 
       20100224__issue16802.diff.txt uploaded by tilghman (license 14)
 Tested by: tzafrir


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@248859 f38db490-d61c-443f-a65b-d21fe96a405b