Remove some hidden broken code in the voicemail mailbox options menu.
After finishing a recording from within the mailbox options menu, pressing 0
exhibited strange behavior with operator=yes turned on. Pressing 0 was not
even advertised as an option and the options from the vm-saveoper prompt:
"Press 1 to accept this recording. Otherwise, please continue to hold" did
not function correctly. While this of course could be fixed, it didn't really
seem to make sense even if it was working properly.

ABE-2121
SWP-1267


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@261698 f38db490-d61c-443f-a65b-d21fe96a405b