Dial and queue connected line update macro not always run when expected.
The connected line update macro would not get run if the connected line
number string was empty.  The number could be empty if the connected line
update did not update a number but the name.  It should be run if there
was an AST_CONTROL_CONNECTED_LINE frame received for pending dials and
queues.

Renamed and added some more comments for some confusing identifiers
directly connected to the related code.

Also fixed a memory leak in app_queue.

Review:	https://reviewboard.asterisk.org/r/669/


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@264669 f38db490-d61c-443f-a65b-d21fe96a405b