Inherit CHANNEL() writes to both sides of a Local channel
Having Local (/n) channels as queue members and setting the language in the
extension with Set(CHANNEL(language)=fr) sets the language on the Local/...,2
channel. Hold time report playbacks happen on the Local/...,1 channel and
therefor do not play in the specified language.

This patch modifies func_channel_write to call the setoption callback and pass
the CHANNEL() write info to the callback. chan_local uses this information to
look up the other side of the channel and apply the same changes to it.

(closes issue #17673)
Reported by: Guggemand

Review: https://reviewboard.asterisk.org/r/903/


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@286059 f38db490-d61c-443f-a65b-d21fe96a405b