Reverted part of r314607, as it can introduce a regression.
Specifically, the security check for the "system" privilege was removed. If a user had the "call" privilege but not the "system" privilege, they would loose the ability to execute the system app and dialplan functions that run commands in a shell. This branch never used the "system" privilege for that purpose and did not need to be patched.

AST-2011-006

(related to issue 0018787)
Reported by: kobaz


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@315147 f38db490-d61c-443f-a65b-d21fe96a405b