Deadlock between generic CCSS agent and native ISDN CCSS.
Deadlock can occur when the generic CCSS agent is deleting duplicate CC
offers and the native ISDN CC driver is processing an incoming CC message.
The cc_core_instances container lock cannot be held when an agent or
monitor callback is invoked without the possibility of a deadlock.

* Make kill_duplicate_offers() remove the reference in cc_core_instances
outside of the container lock.

JIRA AST-566
JIRA SWP-3469


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@319259 f38db490-d61c-443f-a65b-d21fe96a405b