Native SIP CCSS sends bad CC cancel SUBSCRIBE message.
The SUBSCRIBE message used to cancel a CC request has incorrect To/From
SIP headers.  They are reversed and the dialog tags are the same when they
should not be.  If pedantic mode was disabled, then the cancel would have
succeeded despite the incorrect message.

* The SIP_OUTGOING flag was not set correctly for the dialog and I had to
move some CC subscribe handling code as a result.

* Initialized the dialog subscribed type to CALL_COMPLETION earlier.  If a
CC request SUBSCRIBE message comes in and the CC instance is not found,
the 404 response was duplicated.

JIRA AST-568
JIRA SWP-3493


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@320883 f38db490-d61c-443f-a65b-d21fe96a405b