Better way to get chan and pvt lock for issue ASTERISK-17431.
Redoes -r308945 for issue ASTERISK-17431 deadlock fix for
sip_set_udptl_peer() and sip_set_rtp_peer().

* Lock the channels in the defined order and avoid the need for a deadlock
avoidance loop.

* Lock the channel before getting the pointer to the private structure to
be sure that the pointer will not change due to a masquerade or channel
hangup.

* To preserve sanity, check that chan and p->owner are the same.  (Pointer
rearangements should not happen without the protection of locks because
bad things tend to happen otherwise.)


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@326144 f38db490-d61c-443f-a65b-d21fe96a405b