Ensure frames are not written to dialed channel if ringback is requested
When a single channel was dialed and there was media to be forwarded to the
calling channel, the media was written without regard for ringback causing
silence to be heard in some circumstances.  This regression was introduced
when the meaning of "single" changed to mean only the number of channels
dialed.

(closes issue ASTERISK-18083)


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@335341 f38db490-d61c-443f-a65b-d21fe96a405b