Make ast_pbx_run() not default to s@default if extension is not found
Review: https://reviewboard.asterisk.org/r/1446/

This is a bug - or architecture mistake - that has been in Asterisk for a 
very long time. It was exposed by the AMI originate action and possibly
some other applications. Most channel drivers checks if an extension
exists BEFORE starting a pbx on an inbound call, so most calls will
not depend on this issue.

Thanks everyone involved in the review and on IRC and the mailing list
for a quick review and all the feedback.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/10@337219 f38db490-d61c-443f-a65b-d21fe96a405b