Update MeetMe p and X option documentation when interacting with the s option.
ASTERISK-12175 changed the p and X options to not interfere with the s
option when they are used together.  It makes more sense for the s option
to have priority for the DTMF '*' key since it cannot change its
activation code.  Otherwise, you could not use option s with the p or X
options.

JIRA AST-671


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@340470 f38db490-d61c-443f-a65b-d21fe96a405b