In ChanSpy, don't create audiohooks that will never be used.
When ChanSpy is initialized it creates and attaches 3 audiohooks:

  1) Read audio off of the channel that we are spying on
  2) Write audio to the channel that we are spying on
  3) Write audio to the channel that is bridged to the channel that we are
     spying on.

The first is always necessary, but the others are used only when specific
options are passed to the ChanSpy application (B, d, w, and W to be specific).

When those flags are not passed, neither of those audiohooks are ever sent
frames, but we still try to process the hooks for each voice frame that we
recieve on the channel.

So in short - only create and attach audiohooks that we actually need.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@349044 f38db490-d61c-443f-a65b-d21fe96a405b