Fix translation path choices.
This change makes it so computational cost is not taken into account
when deciding if a multistep path is better than a single-step path. This
means that the only time a multistep path will be chosen is if no single-step
path exists. This ensures a better quality translation even if it turns out
to be slightly slower.

(closes issue ASTERISK-16821)
reported by Andrew Lindh

Review: https://reviewboard.asterisk.org/r/1715



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@354594 f38db490-d61c-443f-a65b-d21fe96a405b