Add additional character type types to supported data types for cdr_adaptive_odbc
The reporter was uable to use varchar utf8_unicode_ci with cdr_adaptive_odbc, so
this patch adds those along with some other character types to the list of types
cdr_adaptive_odbc will work using the varchar conditions. The problem wasn't really
UTF8 characters as much as it was a failure to respond to the exact type that was
declared/in use on that database.

(closes issue ASTERISK-19334)
Reported By: Igor Nikolaev
Patches:
	cdr_adaptive_odbc.patch uploaded by Igor Nikolaev (license 6236)




git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@357455 f38db490-d61c-443f-a65b-d21fe96a405b