codec_dahdi: Block on frameout if the hardware has enough samples to complete a frame.
Given the following generic scenario:
RTP(g729) -> codec_dahdi -> RTP(ulaw) -> chan_dahdi -> phones

I received a report that if codec_dahdi was replaced with codec_g729 skipping
audio was no longer observed on the phones on the far left. It appeared that
there were bursts of audio that could cause frames to accumulate at the
ingress to chan_dahdi and potentially the jitter buffer was interfering.

Regardless, the below change, which makes codec_dahdi operate in a more
synchronous fashion, like codec_g729, made the problem go away for the user.
As a side effect, use the file conversion tools from the Asterisk CLI now also
work.

The downside is that the thread responsible for transcoding the audio may be
blocked for several milliseconds while it waits for the hardware transcoder to
complete, but other threads in the system are free to do work and the overall
load on the system remains low.

Internal-Issue-ID: DAHDI-953
Signed-off-by: Shaun Ruffell <sruffell@digium.com>