Clean up after a reinvite that never gets a final response
The basic problem is that if a re-INVITE is sent by Asterisk and it receives a
provisional response, but no final response, then the dialog is never torn
down. In addition to leaking memory, this also leaks file descriptors and will
eventually lead to Asterisk no longer being able to process calls.

This patch just keeps track of whether there is an outstanding re-INVITE, and if
there is goes ahead and cleans up everything as though there was no outstanding
reinvite.

(closes issue ASTERISK-19992)


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@369436 f38db490-d61c-443f-a65b-d21fe96a405b