Free any datastores attached to dummy channels.
Revision 370205 added the use of a datastore attached to a dummy channel to
resolve a memory leak, but ast_dummy_channel_destructor() in this branch did
not free datastores, resulting in a continued (but slightly smaller) memory
leak. This patch backports the change to free said datastores from the Asterisk
trunk.

(related to issue AST-916)


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@370360 f38db490-d61c-443f-a65b-d21fe96a405b