Fix misleading documentation in agents.conf.sample regarding ackcall usage.
The documentation made it sound as if the DTMF acknowledgment was needed
at the time the agent logs in, rather than when the agent is called. This
is likely a relic from the days when there were multiple ways of logging
in agents.

(closes issue AST-962)
reported by Steve Pitts



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@371787 f38db490-d61c-443f-a65b-d21fe96a405b