Free ast_str objects when temp file fails to be created in MiniVM
The previous commit (r372554) was from a patch that was written before
r366880, which ensured that ast_str objects allocated in the sendmail
routine were free'd in off nominal paths.  This commit frees the
string objects in the off nominal path introduced in r372554.

(issue ASTERISK-17133)
Reported by: Tzafrir Cohen


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@372581 f38db490-d61c-443f-a65b-d21fe96a405b