Ensure that all ConfBridge sounds can be set using CONFBRIDGE function
The CONFBRIDGE function can be used to set the sounds in a ConfBridge
bridge profile.  Unfortunately, three sounds were missed in the portion
of the code that applies the settings passed in from the function:
sound_only_one, join, and leave.  This patch makes sure that the sounds
passed from the function are applied to the bridge profile.

(closes issue ASTERISK-20404)
Reported by: univ
Tested by: mjordan


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/10@373196 f38db490-d61c-443f-a65b-d21fe96a405b