Makefile: Add "tags" and "cscope" targets.
When in the Asterisk directly a developer can now run:

  $ make tags cscope

to create both a local tags file and the cscope database. You can
also create either independently. I find that cscope is useful for
some tasks but that ctags does a better job of locating definitions
that have preprocessor macros.  Therefore, they are both useful to
have when browsing through the source code.

Signed-off-by: Shaun Ruffell <sruffell@digium.com>

Conflicts:
	Makefile

Conflicts:
	Makefile