Fix regression in Confbridge user count
When the restructuring work got committed to Confbridge in r375470 to
fix many open issues, it caused a regression in the reported count of
users when conference information was requested via CLI or manager.
This corrects the user count and user information displayed when
listing conference information from the CLI and manager.

(closes issue ASTERISK-20938)
Reported-by: Timo Teras
Patch-by: Timo Teras


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/11@379478 f38db490-d61c-443f-a65b-d21fe96a405b