res_parking: Fix some simple bugs
Both of them are covered in the dynamic parking review on
https://reviewboard.asterisk.org/r/2550 - Remove unref against
parking lot that the bridge did on dissolve since the reference
wasn't taken in the first place. On a swap, reapply bridge roles
in order to get music on hold and such playing on the channel that
swaps into the bridge.


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@389618 f38db490-d61c-443f-a65b-d21fe96a405b