Fix method for creating activities string in PIDF bodies.
The previous method did not allocate enough space to create
the entire string, but adjusted the string's slen value to
be larger than the actual allocation. This resulted in garbled
text in NOTIFY requests from Asterisk.

This method allocates the proper amount of space first and then
writes the content into the buffer.



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/12@397960 f38db490-d61c-443f-a65b-d21fe96a405b