Fix Regression With Queuelog EXITWITHKEY Only Logging Two Out Of Four Fields
Commit r62462 added two extra fields for logging "the original position the 
caller entered the queue at, and the amount of time the caller was waiting in
the queue."  But when r75969 was merged from 1.4 into trunk (r75977), these two
fields disappeared. Those two extra fields were not logged in 1.4 and when the
patch was merged, those fields went away.

Therefore, this is a regression and was caught by the reporter because he was
reading the awesome "Asterisk: The Definitive Guide" book.

(closes issue ASTERISK-22197)
Reported by: Dalius M.
Tested by: Dalius M.
Patches:
    asterisk-22197-q-log-exitwithkey.diff
				     uploaded by Michael L. Young (license 5026)

Review: https://reviewboard.asterisk.org/r/2901/


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@400622 f38db490-d61c-443f-a65b-d21fe96a405b