pjsip outbound registration: Log message says received a 408 when we didn't
If the server didn't exist that we are trying to register to the log message
would say that a 408 was received from that server when in reality one wasn't.
Added log messages stating no response was received if the response does not
exist.

(closes issue ASTERISK-22554)
Reported by: Rusty Newton
Review: https://reviewboard.asterisk.org/r/2893/


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/12@400890 f38db490-d61c-443f-a65b-d21fe96a405b