res_pjsip: Update handling of some options to work with new option names.
Some options (such as call_group and pickup_group) share the same configuration
handler and decide what logic to use based on the name of the option. These
handlers were not updated to check for the new option names and were treating
the options as invalid.

This change simply updates the handlers with the proper names of the options.

(closes issue ASTERISK-22922)
Reported by: Anthony Messina


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/12@403173 f38db490-d61c-443f-a65b-d21fe96a405b