Initialize the hash value argument to pj_hash_get() to 0.
Passing a non-zero value causes PJLIB to use the given input as the
hash value. Passing zero causes the parameter to become an output parameter
that receives the hash value that was computed based on the given key.

This change essentially makes ast_sip_dict_get() properly retrieve the
desired value.



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/12@403349 f38db490-d61c-443f-a65b-d21fe96a405b