Fixup skinny registration following network issues.
On session registration, if device is already reporting that it is
connected to a device, an innocuous packet (update time) is sent to
the already connected device. If the tcp connection is down, the
device will be unregistered and the new connection allowed.

Without this patch, network issues can see a situation where a device
can not reregister until after 3*timeout.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/12@404292 f38db490-d61c-443f-a65b-d21fe96a405b