res/Makefile: alias dist-clean to distclean
A 'make distclean' is equivalent to 'make dist-clean' in the top most Makefile.
This patch updates the res/Makefile to recognize both distclean and dist-clean.
Note that this is needed for removing build.mak, which can run into problems
if the source file of Asterisk or its path is changed after build.mak is
generated.

(issue ASTERISK-22480)
Reported by: Matt Jordan


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/11@405362 f38db490-d61c-443f-a65b-d21fe96a405b