CHANGES: Improved description of Name/Creator changes to bridge ARI, adds AMI
The changes log was written with language that was a little too internal
Asterisk specific, so it's been changed to be more in the frame of reference
of an ARI user. Also, previously the AMI event changes were omitted from the
change log as well as the ability to include a bridge name in the ARI post
bridges command.
........

Merged revisions 407461 from http://svn.asterisk.org/svn/asterisk/branches/12


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@407462 f38db490-d61c-443f-a65b-d21fe96a405b