wip:codec_dahdi: Add configurable "synchronous" mode.
When codec_dahdi is in synchronous mode (the default) it will synchronously transcode each
frame. For example when framein is called, codec_dahdi will block until the transcoded
frame is ready. This can take up to 6ms.  When synchronous mode is disabled, if the
hardware does not have a ready frame, it will return a "silent" frame instead of blocking.
This will increase the delay of the audio through the system but will not block.

Internal-Issue-ID: ASTERISK-23262