Make res_sorcery_realtime filter unknown retrieved results.
When retrieving data from a database or other realtime backend, it's quite
possible to retrieve variables that Asterisk does not care about but that
are legitimate to exist. Asterisk does not need to throw a hissy fit when
these variables are encountered but rather just filter them out.

Review: https://reviewboard.asterisk.org/r/3305



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/12@410187 f38db490-d61c-443f-a65b-d21fe96a405b