http: Fix spurious ERROR message in responses with no content
When a response has a content length of 0, fwrite would be called to write a
buffer with no data in it. This resulted in the following classic error
message:

  [Apr  3 11:49:17] ERROR[26421] http.c: fwrite() failed: Success

This patch makes it so that we only attempt to write out the content if the
calculated content_length is non-zero.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/12@411687 f38db490-d61c-443f-a65b-d21fe96a405b