astobj2: Add an ao2_replace macro to astobj2.h
This macro replaces one object reference with another cleaning up the original.

param dst Pointer to the object that will be cleaned up.
param src Pointer to the object replacing it.

src's ref count is bumped if it's non-NULL.
dst's ref count is decremented if it's non-NULL.
src is assigned to dst,

This patch was reviewed on IRC by coreyfarrell and mjordan.
 
Tested by: George Joseph


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/12@416995 f38db490-d61c-443f-a65b-d21fe96a405b