event.c: Fix type mismatch errors in ie_maps[].
In v12+ the type values from the table are only used by the CEL unit
tests.  Since the unit tests were only comparing a generated expected
event with a real event to see if the ie contents matched and using the
same table IE_PLTYPE values to read the event contents, the type
mismatches were not detected.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/12@417565 f38db490-d61c-443f-a65b-d21fe96a405b