manager: Return ActionID on nominal responses to PresenceState action
When the PresenceState action is executed, the nominal path fails to include
the ActionID in the successful response. This patch adds a call to
astman_start_ack, which guarantees that an ActionID (if provided) will be
sent back to the AMI client.

Review: https://reviewboard.asterisk.org/r/3776/

ASTERISK-23985 #close


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/11@418713 f38db490-d61c-443f-a65b-d21fe96a405b