Fix build when pjproject is installed in a non-standard location.
When configuring Asterisk to build against a version of pjproject installed
in a non-standard location, the checks for "PJSIP Transaction Group Lock
Support" and "PJSIP Media Stream Replacement Support" fail.  This is
because these secondary checks are not taking the CFLAGS and LIBS returned
by the pkg-config check into account.

Review: https://reviewboard.asterisk.org/r/3830


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@419077 f38db490-d61c-443f-a65b-d21fe96a405b