res_hep: Remove disabling of modules
These modules were originally specified as being disabled, as they were
introduced midstream in Asterisk 12. That makes it nicer for folks who are
upgrading to a new release in the middle of Asterisk 12. That's not the case
for Asterisk 13: it's a brand new release. There's no reason to have the
modules disabled by default in that case.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/13@420742 f38db490-d61c-443f-a65b-d21fe96a405b