Add sample configuration for resource lists.
On review /r/3977, it was recommended to note in the
sample configuration about the size limitation for
resource lists. However, since there was no section in
the sample configuration at all for resource list
subscriptions, I decided to make a separate commit
where I have added the necessary sample configuration
as well as the size limitation warning.



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/13@422853 f38db490-d61c-443f-a65b-d21fe96a405b