Add API call to determine if format capability structure is "empty".
Empty here means that there are no formats in the format_cap structure
or the only format in it is the "none" format.

I've added calls to check the emptiness of a format_cap in a few places
in order to short-circuit operations that would otherwise be pointless
as well as to prevent some assertions from being triggered in cases
where channels with no formats are used.



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/13@423414 f38db490-d61c-443f-a65b-d21fe96a405b