res/res_pjsip_sdp_rtp: Check joint caps when looking to decline outgoing media
When constructing an outgoing media stream for an SDP offer/answer, the current
code checks the override preferences (set by the PJSIP_MEDIA_OFFER function) as
well as what is configured on the endpoint to determine if a codec is available
for the media stream. Unfortunately, this isn't good enough: we must also look
at the negotiated (joint) format capabilities. Otherwise, we'll construct a
media stream offer/answer with no codecs.

Note that this isn't an issue in 13, which already looks at the joint
capabilities thanks to the media re-work done there.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/12@425921 f38db490-d61c-443f-a65b-d21fe96a405b