configs/features.conf: Add documentation noting potential chan_agent conflict
In chan_agent, a '*' is used by default to terminate a bridge with a caller.
This can lead to all sorts of problems if '*' is used by a feature in
features.conf, as the chan_agent disconnect '*' may be detected first.

This patch adds a documentation snippet to features.conf so that users who
attempt to use features with agents know of the potential conflict.

ASTERISK-20402 #close
Reported by: Matt Riddell
patches:
  features.conf.diff uploaded by Matt Riddell (License 5023)


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/11@427617 f38db490-d61c-443f-a65b-d21fe96a405b