Allow for transferer to retry when dialing an invalid extension.
This allows for a configurable number of attempts for a transferer
to dial an extension to transfer the call to. For Asterisk 13, the
default values are such that upgrading between versions will not
cause a behaivour change. For trunk, though, the defaults will be
changed to be more user-friendly.

Review: https://reviewboard.asterisk.org/r/4167



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/13@428145 f38db490-d61c-443f-a65b-d21fe96a405b