core: avoid possible asterisk -r crash from long id
When connecting to the remote console, an id string
is first provided that consts of the hostname, pid,
and version.  This is parsed by the remote instance
using a buffer that may be too short, and can allow
a buffer overrun because it is not terminated. This
patch adds termination and a larger buffer.

Review: https://reviewboard.asterisk.org/r/4182/



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/13@429223 f38db490-d61c-443f-a65b-d21fe96a405b