Blocked revisions 430252
........
contrib/ast-db-manage: Correct down_revision path for user_eq_phone

When the user_eq_phone patch was backported to 13, it referenced the downward
revision that the PJSIP optimistic encryption option also references. This
creates a multi-path upgrade Exception when generating the SQL files.

This patch corrects this in the 13 branch. Note that trunk, which already
contained both of these features, is unaffected by this problem.


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@430254 f38db490-d61c-443f-a65b-d21fe96a405b