main/rtp_engine: Format NTP timestamps as unsigned longs
When the RTCP reports are created, the NTP timestamps are stored as strings,
as JSON does not have an integer type long enough to store the value. However,
on 32-bit systems, a signed long may overflow for some portion of the
timestamp.

This patch corrects the overflow by formatting the timestamps as unsigned
longs.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/13@430840 f38db490-d61c-443f-a65b-d21fe96a405b