res_ari_channels: Return a 404 response when a requested channel variable does not exist.
This change makes it so that if a channel variable is requested and it does not exist
a 404 response will be returned instead of an allocation failed response. This makes
it easier to debug and figure out what is going on for a user.

ASTERISK-24677 #close
Reported by: Joshua Colp


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/13@432154 f38db490-d61c-443f-a65b-d21fe96a405b