Blocked revisions 434708
........
main/event: Remove unnecessary assignment of negative value to enum

When cleaning up some clang compiler warnings, the comparison of a negative
value to an unsigned enum was removed. However, the initial assignment of a
negative value to said enum remained in the variable declaration. This patch
removes that assignment.

Thanks to ibercom in #asterisk-bugs for pointing it out.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/13@434709 f38db490-d61c-443f-a65b-d21fe96a405b